refactor(client): decompose AnnotationCard.tsx into 3 sub-components (#403)
Extract AnnotationEditForm, AnnotationCardActions, and ReplyThread from
a 555-LOC monolith so each unit has one clear responsibility. Reply state
(isReplying, replyText, isSendingReply) and handlers move into ReplyThread;
edit state stays in the parent since enterEditMode lives in the header.
All data-testid attributes preserved exactly.
